pyproject.toml
[project]
name = "best-practices-rag"
version = "0.0.1"
description = "A Claude Code skill pipeline for technology best practices via Neo4j knowledge graph"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "exa-py>=2.6.1",
    "llama-index-core>=0.13.0",
    "llama-index-graph-stores-neo4j>=0.5.1",
    "neo4j>=5.28.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
]

[project.scripts]
best-practices-rag = "best_practices_rag.cli:main"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
addopts = "-m 'not integration'"
asyncio_mode = "auto"
filterwarnings = [
    "ignore:directory .* does not exist:UserWarning:pydantic_settings",
]
markers = [
    "integration: marks tests requiring live external dependencies (Neo4j, inference API, Exa)",
    "smoke: marks live smoke tests requiring all external services",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["best_practices_rag"]

[tool.hatch.build.targets.wheel.force-include]
"best_practices_rag/_claude_files" = "best_practices_rag/_claude_files"
